latest anouncements

+++ Modules/Article/Http/Controllers/PostsController.php
@@ -76,7 +76,9 @@ class PostsController extends Controller
     {
         if ($request->hasFile('image')) {
             $old_image = $post->image;
-            unlink($old_image);
+            if (file_exists($old_image)) {
+                unlink($old_image);
+            }
 
             $image = $request->image;
             $destinationPath = 'uploads/';
@@ -84,7 +86,7 @@ class PostsController extends Controller
             $post->image = Helper::uploadFile($destinationPath, $image); //using helper file
             $post->save(); //saving to database
 
-           
+
         }
 
         if ($request->tags) {
@@ -106,7 +108,10 @@ class PostsController extends Controller
     {
         $post = Post::withTrashed()->where('id', $id)->firstOrFail();
         if ($post->trashed()) {
-            Storage::delete($post->image);
+            $old_image = $post->image;
+            if (file_exists($old_image)) {
+                unlink($old_image);
+            }
             $post->forceDelete();
             return redirect(route('trashed-posts.index'));
         } else {
